examples/asyncpg-postgres/requirements.txt
dbop-core[asyncpg]
asyncpg>=0.29.0
python-dotenv
